backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.12.1,<2.0.0",
    "bcrypt==4.0.1",
    "faker>=30.6.0",
    "fastapi[standard]>=0.114.2,<1.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "psycopg[binary]>=3.1.13,<4.0.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "pydantic>2.0",
    "pyjwt>=2.8.0,<3.0.0",
    "sqlmodel>=0.0.21,<1.0.0",
    "tenacity>=8.2.3,<9.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]
